Split openlibrary/core/vendors/
