+++ openlibrary/core/vendors/amazon.py
@@ -24,18 +24,12 @@ from openlibrary.utils.isbn import (
     normalize_isbn,
 )
 
-logger = logging.getLogger("openlibrary.vendors")
+logger = logging.getLogger("openlibrary.vendors.amazon")
 
-BETTERWORLDBOOKS_BASE_URL = 'https://betterworldbooks.com'
-BETTERWORLDBOOKS_API_URL = (
-    'https://products.betterworldbooks.com/service.aspx?IncludeAmazon=True&ItemId='
-)
-affiliate_server_url = None
-BWB_AFFILIATE_LINK = 'http://www.anrdoezrs.net/links/{}/type/dlg/http://www.betterworldbooks.com/-id-%s'.format(
-    h.affiliate_id('betterworldbooks')
-)
 AMAZON_FULL_DATE_RE = re.compile(r'\d{4}-\d\d-\d\d')
 ISBD_UNIT_PUNCT = ' : '  # ISBD cataloging title-unit separator punctuation
+
+affiliate_server_url = None
 
 
 def setup(config):
@@ -226,7 +220,7 @@ class AmazonAPI:
             'url': "https://www.amazon.com/dp/{}/?tag={}".format(
                 product.asin, h.affiliate_id('amazon')
             ),
-            'source_records': ['amazon:%s' % product.asin],
+            'source_records': [f'amazon:{product.asin}'],
             'isbn_10': [product.asin],
             'isbn_13': [isbn_10_to_isbn_13(product.asin)],
             'price': price and price.display_amount,
@@ -400,7 +394,7 @@ def clean_amazon_metadata_for_load(metadata: dict) -> dict:
         conforming_metadata['subtitle'] = subtitle
     # Record original title if some content has been removed (i.e. parentheses)
     if metadata['title'] != conforming_metadata.get('full_title', title):
-        conforming_metadata['notes'] = "Source title: %s" % metadata['title']
+        conforming_metadata['notes'] = f"Source title: {metadata['title']}"
 
     return conforming_metadata
 
@@ -451,89 +445,3 @@ def cached_get_amazon_metadata(*args, **kwargs):
     result = memoized_get_amazon_metadata(*args, **kwargs)
     # if no result, then recache / update this controller's cached value
     return result or memoized_get_amazon_metadata.update(*args, **kwargs)[0]
-
-
-@public
-def get_betterworldbooks_metadata(isbn: str) -> Optional[dict]:
-    """
-    :param str isbn: Unnormalisied ISBN10 or ISBN13
-    :return: Metadata for a single BWB book, currently listed on their catalog, or
-             an error dict.
-    :rtype: dict or None
-    """
-
-    isbn = normalize_isbn(isbn)
-    try:
-        return _get_betterworldbooks_metadata(isbn)
-    except Exception:
-        logger.exception(f"_get_betterworldbooks_metadata({isbn})")
-        return betterworldbooks_fmt(isbn)
-
-
-def _get_betterworldbooks_metadata(isbn: str) -> Optional[dict]:
-    """Returns price and other metadata (currently minimal)
-    for a book currently available on betterworldbooks.com
-
-    :param str isbn: Normalised ISBN10 or ISBN13
-    :return: Metadata for a single BWB book currently listed on their catalog,
-            or an error dict.
-    :rtype: dict or None
-    """
-
-    url = BETTERWORLDBOOKS_API_URL + isbn
-    response = requests.get(url)
-    if response.status_code != requests.codes.ok:
-        return {'error': response.text, 'code': response.status_code}
-    text = response.text
-    new_qty = re.findall("<TotalNew>([0-9]+)</TotalNew>", text)
-    new_price = re.findall(r"<LowestNewPrice>\$([0-9.]+)</LowestNewPrice>", text)
-    used_price = re.findall(r"<LowestUsedPrice>\$([0-9.]+)</LowestUsedPrice>", text)
-    used_qty = re.findall("<TotalUsed>([0-9]+)</TotalUsed>", text)
-    market_price = re.findall(
-        r"<LowestMarketPrice>\$([0-9.]+)</LowestMarketPrice>", text
-    )
-    price = qlt = None
-
-    if used_qty and used_qty[0] and used_qty[0] != '0':
-        price = used_price[0] if used_price else ''
-        qlt = 'used'
-
-    if new_qty and new_qty[0] and new_qty[0] != '0':
-        _price = new_price[0] if new_price else None
-        if _price and (not price or float(_price) < float(price)):
-            price = _price
-            qlt = 'new'
-
-    market_price = ('$' + market_price[0]) if market_price else None
-    return betterworldbooks_fmt(isbn, qlt, price, market_price)
-
-
-def betterworldbooks_fmt(
-    isbn: str,
-    qlt: Optional[str] = None,
-    price: Optional[str] = None,
-    market_price: Optional[list[str]] = None,
-) -> Optional[dict]:
-    """Defines a standard interface for returning bwb price info
-
-    :param str isbn:
-    :param str qlt: Quality of the book, e.g. "new", "used"
-    :param str price: Price of the book as a decimal str, e.g. "4.28"
-    :rtype: dict or None
-    """
-    price_fmt = f"${price} ({qlt})" if price and qlt else None
-    return {
-        'url': BWB_AFFILIATE_LINK % isbn,
-        'isbn': isbn,
-        'market_price': market_price,
-        'price': price_fmt,
-        'price_amt': price,
-        'qlt': qlt,
-    }
-
-
-cached_get_betterworldbooks_metadata = cache.memcache_memoize(
-    _get_betterworldbooks_metadata,
-    "upstream.code._get_betterworldbooks_metadata",
-    timeout=dateutil.HALF_DAY_SECS,
-)
