+++ scripts/affiliate-server
@@ -14,7 +14,6 @@ start affiliate-server as fastcgi:
 start affiliate-server using gunicorn webserver:
 
     ./scripts/affiliate-server openlibrary.yml --gunicorn -b 0.0.0.0:31337
-
 """
 
 import json
@@ -26,24 +25,28 @@ import threading
 import time
 
 import web
-import yaml
 
 import _init_path  # noqa: F401  Imported for its side effect of setting PYTHONPATH
 
 import infogami
 from infogami import config
 from openlibrary.core import stats
-from openlibrary.core.vendors import AmazonAPI
+from openlibrary.core.vendors.amazon import AmazonAPI
 from openlibrary.utils.dateutil import WEEK_SECS
 from openlibrary.utils.isbn import (
-    normalize_isbn, isbn_13_to_isbn_10, isbn_10_to_isbn_13)
+    normalize_isbn,
+    isbn_13_to_isbn_10,
+    isbn_10_to_isbn_13,
+)
 
 logger = logging.getLogger("affiliate-server")
 
+# fmt: off
 urls = (
     '/isbn/([0-9X-]+)', 'Submit',
     '/status', 'Status',
 )
+# fmt: on
 
 API_MAX_ITEMS_PER_CALL = 10
 API_MAX_WAIT_SECONDS = 0.9
@@ -51,41 +54,35 @@ API_MAX_WAIT_SECONDS = 0.9
 web.amazon_queue = queue.Queue()  # a thread-safe multi-producer, multi-consumer queue
 
 
-def process_amazon_batch(isbn_10s):
+def process_amazon_batch(isbn_10s: list[str]) -> None:
     """
-    def process_amazon_batch(isbn_10s: list[str]) -> None:
-
     Call the Amazon API to get the products for a list of isbn_10s and store
     each product in memcache using amazon_product_{isbn_13} as the cache key.
     """
-    logger.info("process_amazon_batch(): {} items".format(len(isbn_10s)))
+    logger.info(f"process_amazon_batch(): {len(isbn_10s)} items")
     try:
         products = web.amazon_api.get_products(isbn_10s, serialize=True)
     except Exception:
-        logger.exception("amazon_api.get_product({}, serialize=True)".format(isbn_10s))
+        logger.exception(f"amazon_api.get_product({isbn_10s}, serialize=True)")
         products = []
     for product in products:
         # Make sure we have an isbn_13 for cache key
         cache_key = (
-            product.get('isbn_13') and product.get('isbn_13')[0] or
-            isbn_10_to_isbn_13(product.get('isbn_10')[0])
+            product.get('isbn_13')
+            and product.get('isbn_13')[0]
+            or isbn_10_to_isbn_13(product.get('isbn_10')[0])
         )
         cache.memcache_cache.set(
-            'amazon_product_%s' % cache_key, product, expires=WEEK_SECS
+            f'amazon_product_{cache_key}', product, expires=WEEK_SECS
         )
 
 
-def seconds_remaining(start_time):
-    """
-    def seconds_remaining(start_time: int) -> int:
-    """
+def seconds_remaining(start_time: int) -> int:
     return max(API_MAX_WAIT_SECONDS - (time.time() - start_time), 0)
 
 
-def amazon_lookup():
+def amazon_lookup() -> None:
     """
-    def amazon_lookup() -> None:
-
     A separate thread of execution that uses the time up to API_MAX_WAIT_SECONDS to
     create a list of isbn_10s that is not larger than API_MAX_ITEMS_PER_CALL and then
     passes them to process_amazon_batch()
@@ -107,20 +104,26 @@ def amazon_lookup():
 
 threading.Thread(target=amazon_lookup).start()
 
+
 class Status:
     def GET(self):
-        return json.dumps({
-            "queue_size": web.amazon_queue.qsize(),
-            "queue": list(web.amazon_queue.queue),
-        })
+        return json.dumps(
+            {
+                "queue_size": web.amazon_queue.qsize(),
+                "queue": list(web.amazon_queue.queue),
+            }
+        )
 
 
 class Submit:
-
     @classmethod
     def unpack_isbn(cls, isbn):
         isbn = normalize_isbn(isbn.replace('-', ''))
-        isbn10 = isbn if len(isbn) == 10 else isbn.startswith('978') and isbn_13_to_isbn_10(isbn)
+        isbn10 = (
+            isbn
+            if len(isbn) == 10
+            else isbn.startswith('978') and isbn_13_to_isbn_10(isbn)
+        )
         isbn13 = isbn if len(isbn) == 13 else isbn_10_to_isbn_13(isbn)
         return isbn10, isbn13
 
@@ -134,14 +137,12 @@ class Submit:
 
         isbn10, isbn13 = self.unpack_isbn(isbn)
         if not isbn10:
-            return json.dumps({
-                "error": "rejected_isbn",
-                "isbn10": isbn10,
-                "isbn13": isbn13
-            })
+            return json.dumps(
+                {"error": "rejected_isbn", "isbn10": isbn10, "isbn13": isbn13}
+            )
 
         # Cache lookup by isbn13. If there's a hit return the product to the caller
-        product = cache.memcache_cache.get('amazon_product_%s' % isbn13)
+        product = cache.memcache_cache.get(f'amazon_product_{isbn13}')
         if product:
             return json.dumps({"status": "success", "hit": product})
 
@@ -156,28 +157,30 @@ class Submit:
 def load_config(configfile):
     infogami.load_config(configfile)
 
-    #if 'fastcgi' in d:
+    # if 'fastcgi' in d:
     #    web.config.fastcgi = d['fastcgi']
 
     web.amazon_api = None
     args = [
         config.amazon_api.get('key'),
         config.amazon_api.get('secret'),
-        config.amazon_api.get('id')
+        config.amazon_api.get('id'),
     ]
     if all(args):
         web.amazon_api = AmazonAPI(*args, throttling=0.9)
         logger.info("AmazonAPI Initialized")
     else:
-        raise RuntimeError("{} is missing required keys.".format(configfile))
+        raise RuntimeError(f"{configfile} is missing required keys.")
 
 
 def init_sentry(app):
     from openlibrary.utils.sentry import Sentry
+
     sentry = Sentry(getattr(config, 'sentry', {}))
     if sentry.enabled:
         sentry.init()
         sentry.bind_to_webpy_app(app)
+
 
 def setup_env():
     # make sure PYTHON_EGG_CACHE is writable
@@ -186,7 +189,9 @@ def setup_env():
     # required when run as fastcgi
     os.environ['REAL_SCRIPT_NAME'] = ""
 
+
 cache = None
+
 
 def start_server():
     sysargs = sys.argv[1:]
@@ -196,15 +201,16 @@ def start_server():
     load_config(configfile)
     global cache
     from openlibrary.core import cache
+
     # sentry could be loaded here
     # init_sentry(app)
 
     sys.argv = [sys.argv[0]] + list(args)
     app.run()
 
+
 def start_gunicorn_server():
-    """Starts the affiliate server using gunicorn server.
-    """
+    """Starts the affiliate server using gunicorn server."""
     from gunicorn.app.base import Application
 
     configfile = sys.argv.pop(1)
@@ -231,10 +237,12 @@ def https_middleware(app):
     Using that value to overwrite wsgi.url_scheme in the WSGI environ,
     which is used by all redirects and other utilities.
     """
+
     def wrapper(environ, start_response):
         if environ.get('HTTP_X_SCHEME') == 'https':
             environ['wsgi.url_scheme'] = 'https'
         return app(environ, start_response)
+
     return wrapper
